+++ app/build.gradle
@@ -57,7 +57,6 @@ android {
 }
 
 dependencies {
-
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
@@ -66,10 +65,11 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.4.1'
     implementation 'androidx.activity:activity-compose:1.4.0'
     implementation "com.google.accompanist:accompanist-systemuicontroller:0.19.0"
-    implementation 'io.github.raamcosta.compose-destinations:core:1.4.4-beta'
-    ksp 'io.github.raamcosta.compose-destinations:ksp:1.4.4-beta'
+    implementation 'io.github.raamcosta.compose-destinations:core:1.5.8-beta'
+    ksp 'io.github.raamcosta.compose-destinations:ksp:1.5.8-beta'
     implementation "com.google.dagger:hilt-android:$hiltVersion"
     kapt "com.google.dagger:hilt-compiler:$hiltVersion"
+    implementation "androidx.hilt:hilt-navigation-compose:1.0.0"
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
