--- app/src/main/java/com/spoelt/dicepoker/core/BottomBarState.kt
@@ -1,5 +1,0 @@
-package com.spoelt.dicepoker.core
-
-enum class BottomBarState {
-    VISIBLE, HIDDEN
-}
