--- app/src/main/java/com/spoelt/dicepoker/core/constants/Constants.kt
@@ -1,9 +1,0 @@
-package com.spoelt.dicepoker.core.constants
-
-const val SPLASH_SCREEN_IMAGE_INITIAL_VALUE = 0f
-const val SPLASH_SCREEN_IMAGE_TARGET_VALUE = 0.9f
-const val SPLASH_SCREEN_ANIMATION_DURATION = 900
-const val SPLASH_SCREEN_INTERPOLATION_TENSION = 3f
-const val SPLASH_SCREEN_DELAY = 1500L
-
-const val LOGO_WIDTH_PERCENTAGE = 0.6f
