--- app/src/main/java/com/spoelt/dicepoker/core/ui/components/VerticalSpacer.kt
@@ -1,12 +1,0 @@
-package com.spoelt.dicepoker.core.ui.components
-
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.Dp
-
-@Composable
-fun VerticalSpacer(height: Dp) {
-    Spacer(modifier = Modifier.height(height = height))
-}
