+++ app/src/main/java/com/spoelt/dicepoker/di/AppModule.kt
@@ -1,0 +1,4 @@
+package com.spoelt.dicepoker.di
+
+object AppModule {
+}
