+++ app/src/main/java/com/spoelt/dicepoker/ui/navigation/BottomBarState.kt
@@ -1,0 +1,5 @@
+package com.spoelt.dicepoker.ui.navigation
+
+enum class BottomBarState {
+    VISIBLE, HIDDEN
+}
