+++ app/src/main/java/com/spoelt/dicepoker/ui/theme/Color.kt
@@ -1,0 +1,25 @@
+@file:Suppress("MagicNumber")
+
+package com.spoelt.dicepoker.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val LightPrimary = Color(0xFF00ADB5)
+val LightOnPrimary = Color(0xFFFFFFFF)
+val LightSecondary = Color(0xFF653239)
+val LightOnSecondary = Color(0xFFFFFFFF)
+val LightBackground = Color(0xFFF4F4F8)
+val LightOnBackground = Color(0xFF0282828)
+val LightSurface = Color(0xFFFFFFFF)
+val LightOnSurface = Color(0xFF00ADB5)
+val LightDisabledButton = Color(0xFFC0C0C0)
+
+val DarkPrimary = Color(0xFF00ADB5)
+val DarkOnPrimary = Color(0xFFFFFFFF)
+val DarkSecondary = Color(0xFF653239)
+val DarkOnSecondary = Color(0xFFFFFFFF)
+val DarkBackground = Color(0xFFF4F4F8)
+val DarkOnBackground = Color(0xFF0282828)
+val DarkSurface = Color(0xFFFFFFFF)
+val DarkOnSurface = Color(0xFF00ADB5)
+val DarkDisabledButton = Color(0xFFC0C0C0)
