+++ app/src/main/java/com/spoelt/dicepoker/ui/theme/Shape.kt
@@ -1,0 +1,20 @@
+@file:Suppress("MagicNumber")
+
+package com.spoelt.dicepoker.ui.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Shapes
+import androidx.compose.ui.unit.dp
+
+val ButtonShape = RoundedCornerShape(50)
+
+val Shapes = Shapes(
+    small = RoundedCornerShape(4.dp),
+    medium = RoundedCornerShape(4.dp),
+    large = RoundedCornerShape(0.dp)
+)
+
+val TextFieldShape = RoundedCornerShape(50)
+
+val ExpandableCardShape = RoundedCornerShape(50)
+val ExpandedCardShape = RoundedCornerShape(25)
